[MLX] Conform trainer, API surface, and CI to backend conventions
Close the gaps found auditing the MLX backend against the jax/torch/
openvino patterns. The trainer now honors the full compile contract:
steps_per_execution>1 runs every packed batch instead of silently
dropping all but the first; jit_compile/run_eagerly wrap the per-batch
step with mx.compile (explicit state threading of trainable/optimizer/
metrics arrays, restored around each call); and DataParallel shards
data automatically in fit via MLXEpochIterator + distribute_data_input,
mirroring JAXEpochIterator.

- core: convert lists of mx arrays via mx.stack (the numpy round-trip
  forced evaluation, illegal under mx.compile tracing)
- rnn: fix unstack (mlx arrays have no .take; use mx.take) and align
  cudnn_ok with the jax/torch named signature
- parity: export scatter/stop_gradient from __init__; categorical
  defaults to int32 like the other backends
- excluded tests: convert to the short openvino-style format (verified
  identical skip set), drop the 27 steps_per_execution exclusions that
  masked the batch-drop bug, fix the grain_datast typo
- CI: add mlx to the nightly matrix and run mlx_data_parallel.py
  (mlx.launch ring, 2 processes) in actions.yml
- format: run shell/format.sh fallout on mlx files that had not been
  through the repo's pinned ruff

integration_tests/mlx_data_parallel.py gains a fit-based check that
ranks converge identically to the manual-shard reference.

+++ keras/src/backend/mlx/random.py
@@ -25,26 +25,20 @@ def normal(shape, mean=0.0, stddev=1.0, dtype=None, seed=None):
 def uniform(shape, minval=0.0, maxval=1.0, dtype=None, seed=None):
     dtype = dtype or floatx()
     key = _get_key(seed)
-    sample = mlx_random.uniform(
-        shape=shape, low=minval, high=maxval, key=key
-    )
+    sample = mlx_random.uniform(shape=shape, low=minval, high=maxval, key=key)
     return sample.astype(_to_mlx_dtype(dtype))
 
 
-def categorical(logits, num_samples, dtype="int64", seed=None):
+def categorical(logits, num_samples, dtype="int32", seed=None):
     key = _get_key(seed)
     logits = convert_to_tensor(logits)
-    output = mlx_random.categorical(
-        logits, num_samples=num_samples, key=key
-    )
+    output = mlx_random.categorical(logits, num_samples=num_samples, key=key)
     return output.astype(_to_mlx_dtype(dtype))
 
 
 def randint(shape, minval, maxval, dtype="int32", seed=None):
     key = _get_key(seed)
-    sample = mlx_random.randint(
-        low=minval, high=maxval, shape=shape, key=key
-    )
+    sample = mlx_random.randint(low=minval, high=maxval, shape=shape, key=key)
     return sample.astype(_to_mlx_dtype(dtype))
 
 
@@ -140,9 +134,7 @@ def binomial(shape, counts, probabilities, dtype=None, seed=None):
     if max_n == 0:
         return mx.zeros(shape, dtype=_to_mlx_dtype(dtype))
     # Generate max_n Bernoulli trials and sum
-    trials = mx.random.uniform(
-        shape=(max_n,) + tuple(shape), key=key
-    )
+    trials = mx.random.uniform(shape=(max_n,) + tuple(shape), key=key)
     successes = (trials < mx.expand_dims(probabilities, 0)).astype(mx.float32)
     # Mask out trials beyond each element's count
     trial_idx = mx.arange(max_n).reshape((-1,) + (1,) * len(shape))
